Move comment author line to separate component
